Git: No Fast-Forward on merge
"No Fast-Forward" merges are a great default when not working in a team;
merging a feature branch into master while keeping the individual
commits is easy. But on a team, or using git as a software update
tool (Homebrew, Doom, etc.), where you often pull other's changes, that
causes a new merge commit, which is bad. As such, it's better to have FF
be the default, and pass in the `--no-ff` option when merging a feature
branch.