Emacs: Don't set unicode font size
Setting the :size of the unicode font was causing it to be locked to
that size. This is not a good idea according to the apropos
documentation on `doom-unicode-font`. Indeed, this was causing
mis-alignment issues in places like magit's log (the graph was not
aligning, making it hard to follow a branch) and minor issues in code
where lines weren't aligning as expected.